+++ mbsimgui/mbsimgui/element.cc
@@ -22,6 +22,7 @@
 #include <cmath>
 #include "element_view.h"
 #include "parameter_view.h"
+#include "parameter.h"
 #include "frame.h"
 #include "contour.h"
 #include "dynamic_system_solver.h"
@@ -35,6 +36,7 @@
 #include <xercesc/dom/DOMLSSerializer.hpp>
 #include <QDir>
 #include <xercesc/dom/DOMProcessingInstruction.hpp>
+#include <boost/range/adaptor/reversed.hpp>
 
 using namespace std;
 using namespace MBXMLUtils;
@@ -176,6 +178,28 @@ namespace MBSimGUI {
         emit mw->getElementView()->model()->dataChanged(containerIndex, containerIndex, { Qt::UserRole });
       }
     }
+  }
+
+  void Element::setParameterCode(const std::string &parName, const std::string &code) {
+    vector<EmbedItemData*> parents = getEmbedItemParents();
+    parents.emplace_back(this);
+    for(auto & parent : boost::adaptors::reverse(parents))
+      for(int j=parent->getNumberOfParameters()-1; j>=0; j--) {
+        auto parameter=parent->getParameter(j);
+        if(parameter->getName().toStdString()==parName) {
+          auto e=parameter->getXMLElement();
+          // remove all existing child nodes
+          while(e->getFirstChild())
+            e->removeChild(e->getFirstChild());
+          // and insert the code as (the only) text node
+          e->insertBefore(e->getOwnerDocument()->createTextNode(X()%code), nullptr);
+          // update parameter tree
+          parameter->updateValue();
+          // update element names
+          MainWindow::updateNameOfCorrespondingElementAndItsChilds(parameter->getParent()->getModelIndex());
+          return;
+        }
+      }
   }
 
 }