+++ mbsimgui/mbsimgui/parameter.h
@@ -63,6 +63,7 @@ namespace MBSimGUI {
       QString getValue() const override { return value; }
       QString getComment() const override { return comment; }
       void setComment(const QString &comment_) { comment = comment_; }
+      void setValue(const QString &v) { value=v; }
       bool getHidden() const { return hidden; }
 //      QString getValue() const override { return MBXMLUtils::E(element)->getFirstTextChild()?QString::fromStdString(MBXMLUtils::X()%MBXMLUtils::E(element)->getFirstTextChild()->getData()):""; }
       virtual xercesc::DOMElement* createXMLElement(xercesc::DOMNode *parent);