+++ mbsimgui/RELEASENOTES.md
@@ -27,6 +27,10 @@ MBSimGUI
 - Frame chaser observer is now available.
 - Motion observer is available (see MBSimControl).
 - Use the Integrator startTime in the 3D-view of mbsimgui. t=0 was used for now to draw the 3D-view.
+- Update element names if a parameter of a element changes to display the correct new element name.
+- Added plotting: model defined element context menu actions can now be added. When clicked the deposited evaluator code is executed.
+  The output is displayed in the Echo Area. The evaluator code can even create dialogs e.g. using PySide2 and create
+  plots using e.g. matplotlib. Setting parameters from the evaluator code is also possible.
 
 OpenMBV
 -------